Tried to move non-generic collections out of System.Runtime
As a result, I needed to move bunch of security APIs out, which is probably a good thing.

Now, that some of the security types are out, I could probably move more of these out too,
but I will try to do it after we decide we want this change.