Added ComAwareEventInfo
This type already exists in CoreFx and so I have to add it for validation tools pass
The type is being added manually (i.e. I changed manually CoreFxAll.cs master file),
as we don't have a real valid implementation in CoreFx.
By "valid" I mean overriding all the members that are abstract in Xamarin.