Merged Xamarin and CoreFx APIs
So far, the tool would take the Xamarin types (Xamarin members of each type), and then
add CoreFx types if they did not exist in Xamarin. This meant that if a type existed in both,
but the CoreFx type had more members, the members would disappear from the contracts.
This in turn caused ApiCompat to fail may types when checking compatibility with existing
CoreFx contracts.

To fix that, I created a merged master API file that contains CoreFx members merged into
Xamarin types. The merged source file was created using a custom tool and then manually
tweaked (for merge conflicts the tool was not able to handle).

All the contracts now compile, but there are many warnings when they are compiled.
The warnings are related to members being hidden (as opposed to overridden).
I will fix these in subsequent commits.